fix: compiler warning

+++ Platforms/OpenGL/OpenGL_Buffer.cpp
@@ -59,14 +59,14 @@ namespace OpenGL
 		int verStride = getVBOStride();
         std::cout << "Stride Count : " << verStride << std::endl;
         std::cout << "### Position" << std::endl;
-		glVertexAttribPointer(0, getMask(BUFFER_MASKPPP), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)0);
+		glVertexAttribPointer(0, getMask(BUFFER_MASKPPP), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(size_t)0);
 		glEnableVertexAttribArray(0);
 
 		if(getMask(BUFFER_MASKRGB))
 		{
             std::cout << "### RGB" << std::endl;
 			uint colOffset = getMask(BUFFER_MASKPPP)*sizeof(float);
-			glVertexAttribPointer(1, ((getMask(BUFFER_MASKRGB))>>3), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(colOffset));
+			glVertexAttribPointer(1, ((getMask(BUFFER_MASKRGB))>>3), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(size_t)(colOffset));
 			glEnableVertexAttribArray(1);
 		}
 		
@@ -74,7 +74,7 @@ namespace OpenGL
 		{
             std::cout << "### UV" << std::endl;
 			uint texOffset = (getMask(BUFFER_MASKPPP)+(getMask(BUFFER_MASKRGB)>>3))*sizeof(float);
-			glVertexAttribPointer(2, ((getMask(BUFFER_MASKTEX)>>6)), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(texOffset));
+			glVertexAttribPointer(2, ((getMask(BUFFER_MASKTEX)>>6)), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(size_t)(texOffset));
 			glEnableVertexAttribArray(2);
 		}
 
@@ -82,7 +82,7 @@ namespace OpenGL
 		{
             std::cout << "### Normal" << std::endl;
 			uint norOffset = (getMask(BUFFER_MASKPPP)+(getMask(BUFFER_MASKRGB)>>3)+(getMask(BUFFER_MASKTEX)>>6))*sizeof(float);
-			glVertexAttribPointer(3, ((getMask(BUFFER_MASKNORMAL)>>9)), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(norOffset));
+			glVertexAttribPointer(3, ((getMask(BUFFER_MASKNORMAL)>>9)), GL_FLOAT, GL_FALSE, sizeof(float)*verStride, (const void*)(size_t)(norOffset));
 			glEnableVertexAttribArray(3);
 		}
     }
